Construct an item to be contained by CSS grid layout

Construct an item to be contained by CSS grid layout

grid_item(..., row_start = NULL, row_end = NULL, row = NULL,
  column_start = NULL, column_end = NULL, column = NULL,
  area = NULL, justify_self = c("", "start", "center", "end"),
  align_self = c("", "start", "center", "end"), z_index = NULL,
  order = NULL, style = NULL)

Arguments

...
Items in the container

row_start, row_end, row, column_start, column_end, column, area
CSS properties of gird-*, whereas grid- is abbreviated, and hyphens are replaced by underscores. For example, the row_start argument is equivalent to the CSS's grid-row-start property. Specify a string as a value for the corresponding CSS property.

justify_self, align_self
One of "", "start", "center", or "end" which can be partial matched. "" does not specify the corresponding CSS property (justify-self, and align-self, respctively).

z_index, order
A string to specify the corresponding CSS property (z-index and order, respectively)

style
Additional values of the style attribute such as color: red;.